Feature - Show Litecoin Card as Webview (#82)
* Feature - Show Litecoin Card as Webview
 - remove partner api library
 - create card frament webview and add to bottom nav
 - fixes #58

* bump version code

* Fix UI crash

* bump version code
