+++ app/build.gradle
@@ -64,8 +64,8 @@ android {
         applicationId = 'com.loafwallet'
         minSdkVersion 27 
         targetSdkVersion 32 
-        versionCode 705
-        versionName "v2.7.1" 
+        versionCode 718
+        versionName "v2.7.1"
         multiDexEnabled true
         archivesBaseName = "${versionName}(${versionCode})"
 
@@ -316,8 +316,6 @@ dependencies {
 
     // Progress Button
     implementation 'com.github.razir.progressbutton:progressbutton:2.1.0'
-
-    api project(":partnerapi")
 
     // Dagger
     implementation 'com.google.dagger:dagger:2.41'
