+++ app/src/main/java/com/breadwallet/presenter/base/BasePresenter.kt
@@ -1,9 +1,5 @@
 package com.breadwallet.presenter.base
 
-import com.breadwallet.di.component.DaggerPresenterComponent
-import com.breadwallet.presenter.spend.LoginPresenter
-import com.breadwallet.presenter.spend.RegisterPresenter
-import com.breadwallet.presenter.transfer.TransferPresenter 
 
 /** Litewallet
  * Created by Mohamed Barry on 6/30/20
@@ -12,18 +8,12 @@ import com.breadwallet.presenter.transfer.TransferPresenter
  */
 abstract class BasePresenter<out V : BaseView>(var view: BaseView?) {
 
-    private val injector = DaggerPresenterComponent.create()
 
     init {
         inject()
     }
 
     private fun inject() {
-        when (this) {
-            is LoginPresenter -> injector.inject(this)
-            is RegisterPresenter -> injector.inject(this)
-            is TransferPresenter -> injector.inject(this) 
-        }
     }
 
     abstract fun subscribe()
